python-server-demo/requirements.txt
Flask
python-gcm
redis